+++ packages/terra-theme-provider/tests/nightwatch/ThemeProviderTests.jsx
@@ -1,0 +1,15 @@
+/* eslint-disable import/no-extraneous-dependencies */
+
+import React from 'react';
+import { Link } from 'react-router';
+
+const ThemeProviderTests = () => (
+  <div>
+    <ul>
+      <li><Link to="/tests/theme-provider-tests/default">ThemeProvider - Default</Link></li>
+      <li><Link to="/tests/theme-provider-tests/theme-switching">ThemeProvider - Theme Switching</Link></li>
+    </ul>
+  </div>
+);
+
+export default ThemeProviderTests;