+++ packages/terra-site/src/examples/modal-manager/ContentContainer.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import AppDelegate from 'terra-app-delegate';
+import Heading from 'terra-heading';
 
 class ContentContainer extends React.Component {
   constructor(props) {
@@ -67,12 +68,9 @@ class ContentContainer extends React.Component {
 
     return (
       <div className="content-container" style={{ height: '100%', padding: '10px' }}>
-        <h2>Content Component</h2>
-        <br />
-        <h4>id: {identifier}</h4>
-        <br />
-        {app && app.releaseFocus ? <h4>Modal has lost focus!</h4> : null }
-        {app && app.requestFocus ? <h4>Modal has gained focus!</h4> : null }
+        <Heading level={2}>Demo Content Component</Heading>
+        {app && app.releaseFocus ? <Heading level={4}>Modal has lost focus!</Heading> : null }
+        {app && app.requestFocus ? <Heading level={4}>Modal has gained focus!</Heading> : null }
         <br />
         <button className="disclose" onClick={this.disclose()}>Disclose</button>
         {identifier === 'root-component' && <button className="disclose-tiny" onClick={this.disclose('tiny')}>Disclose - Tiny</button>}