+++ packages/terra-site/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Added
 * Added ThemeProvider component examples
+* Added new relic script
 
 1.6.0 - (August 15, 2017)
 ------------------