+++ packages/terra-site/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Added
 * Added line-height styles to site-nav
+* Fixed prop-types error for invalid data type in search field examples
 
 ### Changed
 * Updated various examples to use `span` instead of `p` elements in sample content