+++ packages/terra-site/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 ### Added
+* Added line-height styles to site-nav
+
+### Changed
+* Updated various examples to use `span` instead of `p` elements in sample content
+* Added ThemeProvider component examples
 * Added new relic script
 
 1.6.0 - (August 15, 2017)