+++ packages/terra-site/CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 Unreleased
 ----------
 ### Added
+* Added line-height styles to site-nav
+
+### Changed
+* Updated various examples to use `span` instead of `p` elements in sample content
 * Added ThemeProvider component examples
 * Added new relic script